libjob: [cleanup] break up job.c
Problem: job.c is starting to get too long, and contains disjoint
code that doesn't need to be co-located.

Break up job.c topically into several new .c files.
